Fix missing reasons, wrong codes in error messages
Passing a va_list to a function which takes variadic arguments is not
valid; xcw_die was doing this with xcw_fail which was causing xcw_fail
to print garbage instead of proper reasons and error codes. On my system
the output did tend to look plausible, hence why this probably went
unnoticed for so long.

Added a function xcw_vfail which takes a va_list so that xcw_die may
call it correctly.